// src/types/DotaItemParams.ts
export type DotaItemParams = {
    query: string
    search_descriptions: number
    'category_570_Hero[]': string
    'category_570_Slot[]': string
    'category_570_Quality[]': string
    'category_570_Type[]': string
    start: number
    count: number
    sort_column: string
    sort_dir: string
    appid: number
}